% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{class_esrd_outcome}
\alias{class_esrd_outcome}
\title{Label CKD stages or ESRD outcomes.}
\usage{
class_esrd_outcome(
  df,
  col,
  years,
  duration_col,
  prefix = NULL,
  create_years_col = TRUE
)
}
\arguments{
\item{df}{}

\item{col}{}

\item{years}{}

\item{duration_col}{}

\item{prefix}{}

\item{create_years_col}{}
}
\value{
The modified data frame with added label columns.
}
\description{
Label CKD stages or ESRD outcomes.
}
\examples{
# class_esrd_outcome example
# class_esrd_outcome()
}
