% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{perform_conversions}
\alias{perform_conversions}
\title{Perform unit and code conversions used by KFRE helpers.}
\usage{
perform_conversions(
  df,
  reverse = FALSE,
  convert_all = FALSE,
  upcr_col = NULL,
  calcium_col = NULL,
  phosphate_col = NULL,
  albumin_col = NULL
)
}
\arguments{
\item{df}{}

\item{reverse}{}

\item{convert_all}{}

\item{upcr_col}{}

\item{calcium_col}{}

\item{phosphate_col}{}

\item{albumin_col}{}
}
\value{
A data frame with converted columns.
}
\description{
Perform unit and code conversions used by KFRE helpers.
}
\examples{
# perform_conversions example
# perform_conversions()
}
